add a Consumer Factory class
This class adds a layer of abstraction for
handling having to create and then individually
decorate multiple handler functions when multiple
queues are being used.

Instead, the class handles the instantiation and
decoration of each new queue (or route) internally
and externally the functionality can be accessed by
simply decorating the handler functions with the
same class instance method and a "queue name" arg
instead of having to decorate them with the actual
separate instances of the different consumers.